requirements.txt
numpy>=1.26.4
faiss-cpu>=1.8.0
google-genai>=0.5.0
python-dotenv>=1.0.1
